separate geodescriber from geostore
